Preprocess Notice command
Also provides the beginning of pipeline support in the form of an
"eregs_index". This directory (ignored by git) will serve as a staging area so
that phases of the parsing pipeline can be connected via external files. These
files will also be the component used by a conditional build system (so that
not all files need to be regenerated with every run).

This begins to incorporate some of the ideas the CFPB folks are having around
canonical XML files by moving a piece of meta data (the calculated effective
date of a notice) into an attribute of the XML.