Add a dependency graph to eregs_index.
Most parsing steps involve one or more input files and output files. We'd like
to only run the parsing step if the output file is present and the input file
has changed. The dagger library gives us much of this functionality, using
file timestamps to determine order. We keep track of dependencies in a simple
key-value store which is read at initialization time, allowing the
dependencies to persist between commands