Adds `versions` command.
This command reaches out to the Federal Register to find all final rules and
generates at least one Version for each. The output, serialized Versions will
be used as a canonical source for the parser to know which versions are
available for a given regulation.

This is made a bit more tricky by two complications:

1. A single final rule may be split into multiple Versions, should the rule
have multiple effective dates
2. Rules can modify the effective dates of other rules

Whether or not to (re)compute Version is determined by dependencies, where
each Version depends on the XML of the notice which led to it in addition to
the notice XML for notices which delayed this version. If either has changed,
the Version will be recomputed.