Command for generating missing trees via rules
Once the annual editions are processed, we will likely still have holes in our
timeline of regulation versions. This command accounts for those gaps by
processing the text of final rules, applying the changes to the version
immediately preceding, and serializing the result. Not a super large amount of
code here as we're hooking in to existing systems