Refactor eregs_index.Path into eregs_index.Entry
Referring to entries in the index by tuple was becoming too difficult to
cumbersome, so they've all been replaced with an `Entry`. This has a little
DSL so that `Entry(1, 2, 3) / 4 / 5` is equal to `Entry(1, 2, 3, 4, 5)`. This
also allows the parent Entry to be reused, removing some redundancy.

While VersionPath, TreePath, etc. have their respective replacements
(VersionEntry, TreeEntry, etc.), they are now more uniform. While they
continue to encode/decode in objects, they are now more consistent with
regards to iteration (always returning a string).