Command to fetch Section-by-Section analyses
This more or less copy-pastes functionality from the parse_final_rule command.
Currently the notice dict structure contains both types of data, so we will
just reuse that. This is inefficient, however, and ties the processing of SxS
to amendment parsing, etc., which is not ideal. The two data structures will
split sometime soon, so we'll keep this redundancy for now