Diff command to generate diffs between known trees
Creates a base class which the various JSON-related eregs_index.Entry types
extend and uses that to read FrozenNode trees (which are immutable). This
replicates the diff-generating step of build_from, though only writing to the
index.

Diffs will only be regenerated if the trees they rely on are updated.