First swing at automated dependency resolution
We can determine, by looking at the file name of a missing dependency, how to
resolve that dependency. This adds a `DependencyResolver` base class which is
extended by four commands to describe a resolution to a missing annual
edition, SxS file, rule change file, or notice.

Needed to modify `setup.py` to allow for a recursive call when trying to
resolve.