Don't require write destination to be in settings
Previously, all "write" functions required the end user have their settings
configured appropriately. It'd be nicer if that destination information could
be passed in through the command line.

The only known break from previous methodology is that previously, path names
with forward slashes would automatically be converted to backslashes on a
Windows system. This led to an odd corner case if an absolute UNIX path were
present in the settings. With this patch, we no longer perform the conversion.
This should be fine, as anyone on Windows will be defining their paths with a
backslash anyway.