Add XML preprocessor to merge EXTRACTs.
In one year, the XML might look like:
  <EXTRACT>
    <TAG1>....</TAG1>
    <GPOTABLE>...</GPOTABLE>
    <TAG2>...</TAG1>
  </EXTRACT>

and in the next:
  <EXTRACT>
    <TAG1>....</TAG1>
  </EXTRACT>
  <GPOTABLE>...</GPOTABLE>
  <EXTRACT>
    <TAG2>...</TAG1>
  </EXTRACT>

It's safe to say that the table is supposed to be inside the extract; this
preprocessor performs that change. It also applies these preprocessing steps
to annual editions (previously, they had only been applied to notices), and
adds a shortcut for comparing XML trees.