Adds the `compare_to` command.
The goal of this command is to compare your local tree with a remote API. This
is particularly useful for regression tests, to make sure that tweaks to the
parser didn't change the tree/etc. in unexpected ways.

The command calls `write_to` with a temporary directory, then runs through
the files in that directory, comparing each to the same file on the API. If
the file is not present remotely, the user is notified. If the file has a
difference, the user is asked whether or not they want to see the diff.