Split "or" when finding definitions in keyterms.
Often, definitions will contain multiple, synonymous terms, separated by "or".
This splits those phrases apart into their component definitions. A challenge
here is that "or" isn't a good enough signifier (think of phrases like "red"
or green sign"). To accommodate, this only splits the phrase if all of the
subphrases are single words.

This also converts the keyterm definition search to be a straight parse
attempt. As we only care about the beginning of the line, we don't even need
to try at later parts of the string.