Provide a consistent sort order for Version object
We have been relying on a custom key function which looked at the effective
and publication dates for versions. Now that we've ran in to two notices
published on the same date with the same effective date, we need to also take
into consideration the document id (aka version id). This also moves the
sorting logic into the Version class.