Label creation should be consistent
Previously, we assigned index values for markerless paragraphs which were
unique across the entire section. So, if a markerless paragraph had a
markerless subparagraph, it'd come out as `...-p1-p2`. Extract processing was
different, restarting the paragraph index within the extract, leading to, for
example `...-p1-p1`. This logic, where the index restarts at each paragraph
depth, is now applied to the whole section.