Include child order modifications in diffs
Previously, we were depending on the kindness of `regulations-site` to figure
out how to order the children of a node when displaying a diff. Instead of
having the -site try to recreate the correct order, this patch embeds that
information into the diff.

Along the way, this simplifies FrozenNode's object re-use, gives it a few new
(memoized) properties, and gives it a `clone` method for quick duplication.