Create optional_rules module. Add star_new_level
Split optional depth deriving rules from the rest of the rules. This would
only consist of depth_type_inverses except that this changeset also adds a new
rule, star_new_level. The idea behind this rule is that a STARS paragraph
shouldn't have any subparagraphs (as the hierarchy wouldn't be
understandable). While I think this could probably be used across the board,
it results in an error in an interpretations test. Rather than try to derive
the test's origin, I'm punting and only mixing in this rule when we're working
with regtext.