Speedups and symmetry breaking for depths
Rather than have multiple constraints on a sequence of three paragraphs, we'll
use a single constraint which calls multiple functions. This turns out to be
generally faster, as the logic within those constraints is very simple when
compared to the overhead of adding a new constraint.

In addition, this breaks two additional symmetries, particularly around STARS
and MARKERLESS. See the docstrings for examples.