Rework compare_to interface
Previously, `compare_to` required a CFR title and part as it called the
`write_to` function. It made its comparisons by writing all of the JSON output
to a temporary directory and walking that directory for comparisons.

This patch changes that interface so that it compares _any_ arbitrary local
JSON to a remote service. This is useful as it:

1. doesn't need to duplicate all of the JSON files. This isn't a problem for
smaller regs, but could pose issues for larger regs with many versions
2. doesn't restrict us to comparing a single regulation
3. removes the coupling with write_to
4. allows us use glob syntax for defining what to compare