Allow subscripts to follow more letters
Subscripts used to require that the letter immediately preceding the subscript
be a letter or number. We now also accept the closing paren character. More
importantly, this drops support for the "variable" data field. Although that
information could be useful, trying to figure out nested parentheses isn't
worth the effort.

This change is backwards compatible; -site, which renders the data _was_
converting the text "variable_{subscript}" into {{variable}}<sub>{{subscript}}.
Now we're only giving it "_{subscript}" and leaving the "variable" empty.