current_version command + pipeline flag
Often, we don't need the full history of a regulation, particularly if we just
want to kick the tires. This patch accounts for that scenario by adding a
command for building _only_ the current annual edition of a regulation. The
`pipeline` command has been modified to allow this route be chosen over the
existing, history-centric approach.

To achieve this, we fabricate a version id based on the year and CFR part of
the regulation in question. While not associated with a legitimate notice,
this is a bit more error tolerant, as we can ignore errors deriving effective
dates, etc.