Fix bug w/ intro texts AMDPAR; make API safer
The implementation of AMDPARsing for phrases like "the introductory text for
paragraphs (a) and (b)" had a bug stemming from the interface for creating
Paragraph tokens. Until now, these tokens were constructed with a list
parameter, describing the "label" to associate with the token (note that this
label is slightly different than the labels used with trees). The position of
elements within this list gave them special meaning, which leads to the type
of error introduced. This modifies the API to accept keyword arguments as
well, so that authors can express their intent a bit better. It also fixes
some test cases which were using the interface incorrectly (again, no doubt
due to its error-prone interface)