Fix bug when associating amendments w/ changes
@grapesmoker pointed this out months ago, but it's finally bit us. The code
which ties amendments (e.g. "revising paragraph (b)") with the contents of
those amendments (i.e. the XML nodes representing the text) had a bug where
only the _last_ AMDPAR was being used to determine the associated SECTION XML.

This rewrites that chunk of code to group by section XML rather than CFR part,
which resolves the immediate issue. The code in this area's still pretty
messy, but it's a tad cleaner now