Split CFR parsing from internal citations
Previously, there was special case logic for checking if CFR citations, such
as "26 CFR 111.10" and "26 CFR 111.11(a) through (d)" were actually internal
citations. This generalizes that logic so we can use it outside of the context
of trying to find internal citations. To do this, this moves two functions out
of a function scope (single_citation, multiple_citations), renames a few
things, and adds the CFR title to an internal representation of labels. This
last piece required we add some additional conversion steps as a Node's label
doesn't track CFR titles, yet we want citations to have that ability.