Add a QuickSearchable grammar wrapper
Pyparsing's scanString checks every string index for a parsed string. This is
rather slow, particularly when we're expecting _few_ matches in a given search
string. This adds a wrapper for a Pyparsing grammar which builds an initial
regex to reduce the places we are trying to match the grammar. It should _not_
produce false negatives.

In local testing, it reduces the time to parse all of ATF's regs from scratch
by over 50%