Add upper-case roman numerals to depth parsing
Previously, we tied depth parsing to the types of markers found in regtext,
which excludes upper-case roman numerals. As we'd also like to parse the
depths in congressional statutes, upper-case roman numerals need to be added.
This patch does that while continuing to constrain regtext parsing to using
the markers it was before.