Rework external citations to be modular
We're planning to allow plugin authors to add their own external citation
types. To achieve that goal, this patch modularizes the various types of
external citations into distinct classes with a shared abstract base class
(similar to definition finders and XML preprocessors). This removes the
external_citations grammar (in favor or including those grammars in the finder
classes), converts external citations into a search-replace layer (rather than
indicating offsets), and calculates the link URL here (rather than in -site).

This removes "the Act" as a citation, though it does not remove all of the
setup around that.