Add a simple hierarchical paragraph processor
We need to split `NOTE`s into their component paragraphs to generate better
diffs. Notes may have internal structure, using a variety of paragraph
markers. To step towards this end, this creates a SimpleHierarchyProcessor,
which takes an XML node of arbitrary `P` children and attempts to derive their
depths. We've only seen paragraphs marked with periods, parens, and using
lower case, roman, and Arabic numerals, so that's what this'll cover.