Use "order" when compiling notices more often
Previously, we only used the provided order when there was an exact match
(in terms of node labels) between the order and the children. We'd like to get
away from guessing at order as much as possible. To that end, we'll use the
order provided even when the children are a _subset_ of that order (i.e. not
all children have been added yet).