Account for more whitespace
By default Pyparsing only treats " \t\n" as whitespace chars. Unfortunately,
we've ran across a notice with a "thin-space" unicode character. This patch
configures Pyparsing to account for multiple unicode whitespace chars.

An alternative would be to preprocess all XML and replace thin spaces (and the
like) with normal spaces. We'll see once this rule is released in an annual
edition if that's the correct approach.