Remove a special case for sections/subparts
The compiler had some logic that effectively said "are you a section?" and if
so followed a specific branch of execution separate from the rest of the types
of nodes. We can remove that by giving the `add_node` method a `parent_label`
parameter. This is necessary, particularly for children of subpart and subject
groups, which do not embed their parent.