Refactor constraints on pairs of paragraph markers
Move all constraints on pairs of paragraph markers into a single module; give
them all the same interface and explain their nature. Also use a named tuple
to make it clear what's being compared.

This is a bit more code in terms of number of lines, but should be easier to
test and reason about.