Be more cautious when finding paragraph markers
Originally, we used each Node's label to help find its paragraph markers. In a
recent changeset, this behavior was modified to be much broader, completely
ignoring the label. That turned out to be a mistake, catching lots of
markerless paragraphs.

This patch brings back the check based on the Node's label while still
allowing spans of markers, e.g. "(b) - (d)", be caught.