Replace confusing collapsed marker logic
The grammar we were using to describe collapsed markers was rather difficult
to read. This patch replaces it with the re-usable `any_depth_p`. This means a
bit of the logic which _was_ setting up the grammar (specifically, to only
check for (a), (1), (i), etc.) has been moved to a later step.