Move logic checking for invalid effective date
We'd like to reuse the existing notice preprocessing logic. Unfortunately, it
will blow up when encountering a notice without an effective date. As the end
goal is around proposed rules (which have no effective date), this patch moves
the explosion later in processing, to where the `versions` command is ran