Expand which files write_to looks at
Previously, the `write_to` command would look for all known versions
associated with a particular CFR title/part. It would then use this version
list to determine what information to write.

This patch removes the reliance on versions and the reliance on CFR
title/parts in general. Instead, both title and part are optional arguments
which effectively _filter_ which files are written. Now, _any_ constructed
version, tree, layer, etc. data which is associated with the requested
title/part is sent. If neither filter is given, we'll send everything