Use a better date for eregs current_version
`eregs current_version`, which is used by `pipeline --only-latest` had been
using the current date when creating a version associated with the annual
edition XML. This patch changes the logic to use the publication date of the
annual edition. While this isn't accurate (it's very unlikely that the last
final rule became effective on the date that the annual edition was
published), it's _consistent_ and safer than implying we've captured final
rules which have become effective after the annual edition's publication.