Add warnings for unexpected tags
In the paragraph processor, we were previously ignoring tags which we did not
know how to handle. That's a bit error prone, so this patch adds warnings for
those skipped tags. Further, as there are certain tags we _explicitly_ want to
ignore (e.g. PRTPAGE, the page break marker), we need a way to silence these
warnings for some tags