Allow seq of markerless before markers
We had logic that prevents MARKERLESS and paragraphs with markers from sharing
a parent. For example:

* (a) Contents
* Without marker

became

(a) Contents
  Without marker

This also forced

* Without marker
* More without
* (a) Something

to become

Without marker
More without
  (a) Something

Which isn't quite right. This patch changes the logic so that MARKERLESS and
normal paragraphs can be on the same level so long as the MARKERLESS
paragraphs _come first_.