Replace datatree with simpler solution
While the DSL datatree provides is helpful, the library isn't Python3
compatible and we don't need the bulk of its flexibility. This patch replaces
it with a custom solution. It's backwards compatible (which leads to a rough
edge around the "_xml" attribute), but we'll be able to completely replace
`tests.xml_builder` soon, too.