Add serialization step for Amendment processing
We want to be able to modify the instructions of an AMDPAR without actually
modifying the text. Further, we want the modifications to be very clear to
future devs. To do this, we'll add an XML element to the AMDPAR which spells
out exactly what's changing.

This is a rough implementation which just inserts a
serialization/deserialization step where before there was none.