Delete several unused commands, functions, objects
In an effort to find all of the entry points around building a notice, I found
a very long thread that, when pulled, deleted a lot of unused code. Notably,
we had two deprecated commands, "watch_node" and "notice_order" that continued
to operate outside of the general "pipeline" workflow. As those are no longer
used (and when used, would be inaccurate), they've just been removed. All
supporting functions which only they used have also been removed, notably
including the "Builder" structure, which we've kept around in hopes that we
would _replace_ rather than delete the commands.