Move HTTP cache into index; clear it
Previously, we had the HTTP cache living separately from the rest of the
partially-built content (i.e. the index). This patch moves it into that index
to keep all of the cache content in one directory. This also modifies the
`clear` command, as there's no longer a need for a special flag for clearing
the HTTP cache.