Build with Docker
Containerization should make installation of the parser (a long-standing pain
point) significantly easier. The only major caveat here is that the parser
makes heavy use of caching and allows the user to write results to the local
filesystem. To recreate those features, this recommends creating a shell
script to call docker with flags that'll configure it to meet those needs.