Remove RegPatches custom patching of rules
We had the ability to patch arbitrary changes into the compiled rules. The
ability lead to significant issues as yet another piece of configuration data
had to be shared between developers. As we've since changed the format of the
amendments, it seems like a good time to drop this ability altogether. It'd be
relatively easy to re-add it if we ever need it, but hopefully we won't.