Add amendments to NoticeXML; remove RuleChanges
We had a separate data type for "RuleChanges", that is the collection of
amendments/changes present in a rule. As we're trying to encode all of that
within the XML itself, this adds a property to the NoticeXML wrapper which
fetches the appropriate amendment data. With that, we no longer need
RuleChanges, nor do we need the associated command