Write preambles before writing layers
We ran into an issue a bit before the most recent demo where layer data wasn't
available for the preamble on the new live site. The data _was_ available
locally. Unfortunately, the order-of-operations in the `write_to` command lead
the layers to be written before the preamble; the API needs to look up the
associated document (i.e. the preamble) when processing the layers, so no
layers were added.

This swaps the order so that trees (regulations and preambles) are written
before their layers