Standardize "derive" methods on NoticeXML
The "derive" methods had sometimes stored their value on the NoticeXML object,
sometimes not. Some of the "derive" methods accepted a parameter to set, some
did not. This standardizes so that the "derive" method _only_ read from the
underlying XML. All of the writes have been moved to assignments. I punted on
`derive_agencies` (renaming it `set_agencies`) as I wasn't sure how to handle
it.

The net benefit here is that both `preprocess_notice` and `import_notice` call
`derive_where_needed()`, which will only attempt to derive values which aren't
already set (either from the notice meta data or manual edits)