Rework and optimize regs.gov document retrieval
Rather than make a separate API call for each document type, we'll fetch all
and filter. So long as the official docs don't number more than 1000 (which
seems reasonable), this won't require any paging. Remove the `href` field from
the `RegsGovDoc` class, as it can be easily derived. Add a field for the
"primary" docket (it seems the docket data is messy; we determine primary by
matching on the Federal Register doc id).