Rewrite KeyTerms layer
The goal of this layer is to match any <E T="03"> tags at the beginning of the
node text, ignoring the paragraph marker. It previously performed this by
running multiple passes over the first <E T="03"> in the text via multiple
support methods if `if-then`s. This rewrite tries to simplify by embedding the
checks in the regex (which also simplifies the regex) and reducing the number
of methods.

Also adds a `tagged_text` parameter to `Node`'s constructor.