Set tagged_text when creating table nodes
We've got an ugly mix of relying on `source_xml` and `tagged_text` for
different things throughout this code base. Right now, the formatting layer
assumes that table information will be present in `source_xml`, yet the
notice-compiling code is only aware of `tagged_text`.

This updates the xml-to-Node code for tables to _also_ add `tagged_text`