Use a (Django) setting for index location
We want to be able to configure how the index is stored using helpful pieces
from Django. To help us hit that goal, stop relying on a constant in
regparser.index and prefer sourcing this from Django's settings. Along the
way, add some settings for REQUESTS_CACHE, which we'll make use of shortly