Replace Entry's iteration with sub_entries()
When each Entry had been backed by a file on disk, it made sense to iterate
over them and list sub-directories, contained files, etc. Sub-directories are
no longer present in the db-backed solution. This patch replaces that
iteration with a sub_entries() function which returns full Entries (rather
than strings representing sub-dirs and file names). It also modifies _most_ of
the calling functions to use that.