Use test data from this repo rather than s3
For integration tests, we were using an S3 bucket to read from and store
previously parsed documents. This PR replaces that with _local_ output. This
means we don't have to have the cloud.gov-based S3 bucket. It does require
that PR submitters account fix any discrepancies in their PR, rather than a
merge fixing it.

H/T to @anthonygarvan for describing this option many moons ago.