Host _all_ layers via plugins, start interpparser
The layer configuration system grew up before we had proper extension
management, so we can replace most of it with `stevedore`. Notably, we don't
need to define a `LAYERS` setting, we don't need to worry about layer
shorthands, and we can segment the doc_types by namespace.

This also removes the ALL key, as duplicating the layers explicitly
makes it more obvious what's going on.

Further, this creates a separate interpparser package, installed by default
and converts the relevant tests to use py.test style