Rework layer plugins to be more direct
The last go-around handled layer extensions by creating an extension
point which referenced a _dict_, which contained keys for each doc_type
and lists of class paths as values. This system grew up before we had
proper extension management, so we can replace most of it with
`stevedore`. Notably, we don't need to define a `LAYERS` setting, we
don't need to worry about layer shorthands, and we can segment the
doc_types by namespace.

This also removes the ALL key, as duplicating the layers explicitly
makes it more obvious what's going on.