Account for subpart adds with multiple paragraphs
Previously, we had tacked on the "parent_label" to _all_ subnodes of a subpart
addition. This led all of the nodes to be added at the section level, which
would not only be inaccurate, but lead to an explosion as we couldn't compare
sections (ints) with paragraphs (strings) when sorting.