Replace Version with CFRVersion (backed by the db).
This is step two to moving dependency management into the database -- we now
have SourceFiles in use and CFRVersions being saved/retrieved. We need to
move all of the data from index.entrys into web.index.models, then we can
rework the dependency logic to rely solely on foreign keys.