Replace eregs.py with separate module.
The eregs.py module wasn't included in the setup script, which meant that the
`eregs` command would not work unless run in the same directory. Further,
eregs.py referred to manage.py, which is too generically named to be included
in a package. Instead, move its logic into a module deeper in the app and
reference that module in both the manage.py and `eregs` command. This also
allows us to remove the eregs.py script.